Restructure files
Move helpers.jl and analytic_functions.jl into new folder mathematical_functions.
Move zeeman_shift methods from ion_queries.jl to zeeman_shift.jl.
Move matrix_element, transitionfrequency, and transitionwavelength methods from ion_queries.jl to transitions.jl.
Reclassify energy and quantumnumbers under levels.jl.
Move constants into its own folder, split into mathematical_constants.jl and physical_constants.jl.
Consolidate convenience_functions.

Internal dependency problem -- IonSim does not compile.
