--- src/convenience_functions/_include_convenience_functions.jl
@@ -1,2 +1,0 @@
-include("ion_queries.jl")
-include("set_trap_parameters.jl")
