+++ src/ion/_include_ion.jl
@@ -2,7 +2,6 @@ include("ion_properties.jl")
 include("ion.jl")
 include("ion_instance.jl")
 include("levels.jl")
-include("quantum_numbers.jl")
 include("stark_shift.jl")
 include("zeeman_shift.jl")
 include("transitions.jl")
