+++ src/ion/levels.jl
@@ -7,8 +7,8 @@ export validatesublevel,
     sublevelalias,
     alias2sublevel,
     sublevel2level,
-    leveltransitions,
-    subleveltransitions
+    energy,
+    quantumnumbers
 
 """
     levels(I::Ion)
@@ -126,42 +126,83 @@ function sublevel2level(I::Ion, alias::String)
     return sublevel[1]
 end
 
+# This function is written to be able to accept either a level or sublevel in the second argument
+# Since both levels and aliases are strings, multidispatach can't tell the difference, so the second method distinguishes these cases with an if statement.
 """
-    leveltransitions(I::Ion)
-Returns all allowed transitions between levels of `I` as a vector of `Tuple{String,String}`.
+    energy(I::Ion, sublevel; B=0, ignore_starkshift=false)
+Returns energy of `sublevel` of `I`. A Zeeman shift may be included by setting the value of the magnetic field `B`. The Stark shift may be omitted by setting `ignore_starkshift=true`.
 """
-function leveltransitions(I::Ion)
-    list = []
-    lvls = levels(I)
-    for levelpair in keys(speciesproperties(I).full_transitions)
-        if levelpair[1] in lvls && levelpair[2] in lvls
-            push!(list, levelpair)
-        end
-    end
-    return list
+function energy(I::Ion, sublevel::Tuple{String, Real}; B = 0, ignore_starkshift = false)
+    validatesublevel(I, sublevel)
+    E0 = speciesproperties(I).full_level_structure[sublevel[1]].E
+    zeeman = zeeman_shift(I, sublevel, B)
+    stark = (ignore_starkshift ? 0.0 : stark_shift(I, sublevel))
+    return E0 + zeeman + stark
 end
 
 """
-    subleveltransitions(I::Ion)
-Returns all allowed transitions between sublevels of `I` as a vector of `Tuple{S,S}` where `S=Tuple{String,Real}`.
+    energy(I::Ion, level::trSing)
+Returns the energy of `level` of `I`.
 """
-function subleveltransitions(I::Ion)
-    list = []
-    for transition in leveltransitions(I)
-        (L1, L2) = transition
-        multipole = transitionmultipole(I, L1, L2)
-        sublevels1 = [sublevel for sublevel in sublevels(I) if sublevel[1] == L1]
-        sublevels2 = [sublevel for sublevel in sublevels(I) if sublevel[1] == L2]
-        for sl1 in sublevels1
-            for sl2 in sublevels2
-                m1 = sl1[2]
-                m2 = sl2[2]
-                if abs(m2 - m1) <= parse(Int, multipole[2])
-                    # Only add to list of sublevel transitions if Δm is not larger than the transition multipole allows (1 for E1, 2 for E2, etc)
-                    push!(list, (sl1, sl2))
-                end
-            end
-        end
+function energy(I::Ion, level_or_alias::String; B = 0, ignore_starkshift = false)
+    # If the second argument is a String, it could be either a level name or the alias of a sublevel
+    if level_or_alias in levels(I)
+        # Second argument is a level name. Return the bare energy of that level.
+        return speciesproperties(I).full_level_structure[level_or_alias].E
+    else
+        # Second argument is a sublevel alias.
+        return energy(
+            I,
+            alias2sublevel(I, level_or_alias),
+            B = B,
+            ignore_starkshift = ignore_starkshift
+        )
     end
-    return list
+end
+
+# quantumnumbers is written to be able to accept either a level or sublevel in the second argument
+# Since both levels and aliases are strings, multidispatach can't tell the difference, so the second method distinguishes these cases with an if statement.
+"""
+    quantumnumbers(I::Ion, level::String)
+    quantumnumbers(I::Ion, sublevel)
+Returns the quantum numbers of an energy level or sublevel of `I` as a `NamedTuple`.
+
+If second argument is a level, returns `(:n, :i, :s, :l, :j, :f)`
+
+If second argument is a sublevel, returns `(:n, :i, :s, :l, :j, :f, :m)`
+"""
+function quantumnumbers(I::Ion, sublevel::Tuple{String, Real})
+    validatesublevel(I, sublevel)
+    levelstruct = speciesproperties(I).full_level_structure[sublevel[1]]
+    names = (:n, :i, :s, :l, :j, :f, :m)
+    values = (
+        levelstruct.n,
+        nuclearspin(I),
+        1 // 2,
+        levelstruct.l,
+        levelstruct.j,
+        levelstruct.f,
+        Rational(sublevel[2])
+    )
+    return NamedTuple{names}(values)
+end
+function quantumnumbers(I::Ion, level_or_alias::String)
+    # If the second argument is a String, it could be either a level name or the alias of a sublevel
+    if level_or_alias in levels(I)
+        # Second argument is a level name. Leave out the m quantum number
+        levelstruct = speciesproperties(I).full_level_structure[level_or_alias]
+        names = (:n, :i, :s, :l, :j, :f)
+        values = (
+            levelstruct.n,
+            nuclearspin(I),
+            1 // 2,
+            levelstruct.l,
+            levelstruct.j,
+            levelstruct.f
+        )
+        return NamedTuple{names}(values)
+    else
+        # Second argument is a sublevel alias.
+        quantumnumbers(I, alias2sublevel(I, level_or_alias))
+    end
 end
