+++ src/mathematical_functions/_include_mathematical_functions.jl
@@ -1,0 +1,4 @@
+include("helpers.jl")
+module analytical
+include("analytic_functions.jl")
+end
