+++ src/mathematical_functions/helpers.jl
@@ -72,6 +72,19 @@ function _Dtrunc(Ω, Δ, η, ν, rs, s, n, prefactor, timescale, L, t)
     return g * d, g * conj(d)
 end
 
+function _laguerre(x, n)
+    L = 1.0, -x + 1
+    if n < 2
+        return L[n + 1]
+    end
+    for i in 2:n
+        L = L[2], ((2i - 1 - x) * L[2] - (i - 1) * L[1]) / i
+    end
+    return L[2]
+end
+
+_Pn(n̄::Real, n::Int) = (n̄ / (n̄ + 1))^n / (n̄ + 1)
+
 # associated Laguerre polynomial
 function _alaguerre(x::Real, n::Int, k::Int)
     L = 1.0, -x + k + 1
